refactor: rename i18n directories
